Prevent pvp arena worlds from unlocking new items
A pretty common unlock that people will want is an imbue scroll, and currently when you join a fight in a pvp arena world with the group bronzeman plugin enabled, the items from the pvp arena will be unlocked for the group. These changes should prevent pvp arena worlds from unlocking new items.